major refactor
* Cryptography module with zkps, commitment and elgamal
* zkps module contains decryption and unit_vector zkps
* moved code in `lib.rs` to `tally.rs`
