+++ chain-vote/src/cryptography/zkps/mod.rs
@@ -5,10 +5,9 @@
 //! [spec](https://github.com/input-output-hk/treasury-crypto/blob/master/docs/voting_protocol_spec/Treasury_voting_protocol_spec.pdf),
 //! written by Dmytro Kaidalov.
 
-pub(crate) mod challenge_context;
-pub(crate) mod messages;
-pub mod unit_vector_zkp;
+mod challenge_context;
+mod decr_nizk;
+mod messages;
+mod unit_vector_zkp;
 
-pub(crate) use self::{
-    challenge_context::ChallengeContext, messages::Announcement, unit_vector_zkp::Proof,
-};
+pub(crate) use self::{decr_nizk::ProofDecrypt, unit_vector_zkp::VoteProof};
